refactor(server/players): Take a different approach to multichar
This changes how multicharacter support is implemented entirely.

Instead there are now events for creating a new player & unloading a player. These should be called whenever a character is switched. Documentation will be added for `npwd:newPlayer` & `npwd:unloadPlayer`
