+++ resources/server/functions.ts
@@ -1,7 +1,8 @@
 import { pool } from './db';
 import { config } from './server';
 import { mainLogger } from './sv_logger';
-import { Player, Players, PlayersByIdentifier } from './players/sv_players';
+import { Players, PlayersByIdentifier } from './players/player.provider';
+import { Player } from './players/player.class';
 
 export function getPlayer(source: number): Player {
   const player = Players.get(source);
@@ -35,7 +36,7 @@ export async function getIdentifierByPhoneNumber(
   }
   // Whether we fetch from database if not found in online players
   if (fetch) {
-    const query = `SELECT identifier FROM users WHERE phone_number = ?`;
+    const query = `SELECT ${config.database.identifierColumn} FROM ${config.database.playerTable} WHERE phone_number = ?`;
     const [results] = await pool.query(query, [phoneNumber]);
     // Get identifier from results
     return (results as { identifier: string }[])[0].identifier;
@@ -71,26 +72,33 @@ function getRandomPhoneNumber() {
 }
 
 export async function generatePhoneNumber(identifier: string): Promise<string> {
-  const getQuery = `SELECT phone_number FROM users WHERE identifier = ?`;
+  const getQuery = `SELECT phone_number FROM ${config.database.playerTable} WHERE ${config.database.identifierColumn} = ?`;
   const [results] = await pool.query(getQuery, [identifier]);
   const result = <any[]>results;
   const phoneNumber = result[0]?.phone_number;
 
-  if (!phoneNumber) {
-    let existingId;
-    let newNumber;
-    do {
-      newNumber = getRandomPhoneNumber();
-      try {
-        existingId = await getIdentifierByPhoneNumber(newNumber);
-      } catch (e) {
-        existingId = false;
-      }
-    } while (existingId);
-    mainLogger.debug(`Inserting number into Database: ${newNumber}`);
-    const query = 'UPDATE users SET phone_number = ? WHERE identifier = ?';
-    await pool.query(query, [newNumber, identifier]);
-    return newNumber;
-  }
+  // Generate a new phone number if there isn't one present
+  if (!phoneNumber)
+    if (!phoneNumber) {
+      let existingId;
+      let newNumber;
+
+      do {
+        newNumber = getRandomPhoneNumber();
+        try {
+          existingId = await getIdentifierByPhoneNumber(newNumber);
+        } catch (e) {
+          existingId = false;
+        }
+      } while (existingId);
+
+      mainLogger.debug(
+        `Generated a new number for identifier: ${identifier}, number: ${phoneNumber}`,
+      );
+      const query = `UPDATE ${config.database.playerTable} SET phone_number = ? WHERE ${config.database.identifierColumn} = ?`;
+      await pool.query(query, [newNumber, identifier]);
+      return newNumber;
+    }
+
   return phoneNumber;
 }
