+++ resources/server/players/player.interfaces.ts
@@ -1,0 +1,13 @@
+export interface PlayerAddData {
+  source: number;
+  identifier: string;
+  firstname?: string;
+  lastname?: string;
+}
+
+export interface CreatePlayerInstance {
+  source: number;
+  username: string;
+  identifier: string;
+  phoneNumber: string;
+}
