+++ resources/server/players/player.provider.ts
@@ -1,0 +1,6 @@
+// Map of players by source
+import { Player } from './player.class';
+
+export const Players = new Map<number, Player>();
+// Map of players by identifier
+export const PlayersByIdentifier = new Map<string, Player>();
